safer setup.sh

+++ setup.sh
@@ -1,1 +1,5 @@
-cp config.sample.js config.js
+#!/bin/bash
+
+if [ ! -f config.js]; then
+  cp config.sample.js config.js
+fi
